==> src/collectCvePatches/src/notebooks/nb2.ipynb <==
from typing import List

import pandas as pd
from pathlib import Path
import os
import sys
from dotenv import load_dotenv, dotenv_values
from pydriller import Repository
import json
from multiprocessing import Pool
from itertools import chain

load_dotenv()
ENV = dotenv_values(".env")
DATA_DIR = Path(ENV["DATA_DIR"])
DATA_DIR

df = pd.read_csv(DATA_DIR / 'interim' / 'interim_result.1.tsv', delimiter='\t')
df

df_dict = df.to_dict(orient='records')

def get_new_records(r) -> List:
    new_records = list()
    repo_path = DATA_DIR / 'interim' / 'repositories'
    repo_path.mkdir(exist_ok=True, parents=True)

    try:
        repo = Repository(r["fixCommitRepo"], single=r["fixCommitHash"], clone_repo_to=repo_path)
    except Exception:
        print(f'error cloning {r["fixCommitRepo"]} ({r["fixCommitHash"]})')
        return []

    try:
        commit = list(repo.traverse_commits())[0]
    except Exception:
        print(f'error getting commit {r["fixCommitRepo"]} ({r["fixCommitHash"]})')
        return []

    # assert commit.hash == r["fixCommitHash"]
    for mf in commit.modified_files:
        try:
            if not mf.filename.endswith('.java') or mf.source_code_before is None:
                continue
            new_row = r.copy()
            new_row['commitHashFromGit'] = commit.hash
            new_row['modifiedFilePathBefore'] = mf.old_path
            new_row['modifiedFilePathAfter'] = mf.new_path
            new_row['modifiedFileSrcBefore'] = mf.source_code_before
            new_row['modifiedFileSrcAfter'] = mf.source_code
            new_row['diffParsedJson'] = json.dumps(mf.diff_parsed)
            new_row['nloc'] = mf.nloc
            new_row['changedMethods'] = list(map(lambda m: m.name, mf.changed_methods))
        
            new_records.append(new_row)

        except Exception:
            print(f'error parsing modified files {r["fixCommitRepo"]} ({r["fixCommitHash"]} ({mf.filename})')
            continue

    return new_records


with Pool(8) as p:
    new_records_list = p.map(get_new_records, df_dict)
    new_df = list(chain(*new_records_list))

pd.DataFrame(data=new_df)

last output:
error getting commit https://github.com/apache/poi (3a328aa220f6979f9805f658ae33244d153beaa7)
error getting commit http://svn.apache.org/repos/asf/xerces/java (787352)
error getting commit http://svn.apache.org/repos/asf/activemq (1209700)
error getting commit http://svn.apache.org/repos/asf/httpcomponents/httpclient (1074473)
error getting commit https://github.com/bcgit/bc-java (81b00861cd5711e85fe8dce2a0e119f684120255)

df2 = pd.DataFrame(data=new_df)
df2 = df2[~df2['modifiedFilePathAfter'].isnull()]
df2 = df2[~df2['modifiedFileSrcAfter'].isnull()]
df2

For 11 records, the file has been deleted.

df2 = df2[~df2['modifiedFilePathBefore'].isnull()]
df2 = df2[~df2['modifiedFileSrcBefore'].isnull()]
df2

For 0 records, new file has been introduced.
639 records have been saved

df2.to_json(DATA_DIR/'interim'/'changes_from_git.json', orient='records', indent=2)

df2.columns.tolist()
