==> src/collectCvePatches/src/notebooks/nb6.ipynb <==
from typing import List

import pandas as pd
from pathlib import Path
import os
import sys
from dotenv import load_dotenv, dotenv_values
from pydriller import Repository
import json
from multiprocessing import Pool
from itertools import chain
from collections import OrderedDict
import shutil

load_dotenv()
ENV = dotenv_values(".env")
DATA_DIR = Path(ENV["DATA_DIR"])
DATA_DIR

df: pd.DataFrame = pd.read_csv(DATA_DIR / 'proc' / 'changes.tsv', delimiter='\t')
df

print('\n'.join(df.columns))

msr_dir = DATA_DIR / 'proc' / 'msr24'
jars_dir = msr_dir / 'jars' / 'NEQ3'
jars_dir.mkdir(parents=True, exist_ok=True)

changes_dir = DATA_DIR / 'proc' / 'changes'

df_msr: list[OrderedDict] = []
for it in df.to_dict(orient='records'):
    r = OrderedDict()

    jar_before = jars_dir / Path(it['jarBefore']).name
    jar_after = jars_dir / Path(it['jarAfter']).name

    if not jar_before.is_file():
        shutil.copyfile(DATA_DIR / Path(it['jarBefore']), jar_before)

    if not jar_after.is_file():
        shutil.copyfile(DATA_DIR / Path(it['jarAfter']), jar_after)
    
    r['container_1'] = jar_before.relative_to(msr_dir).__str__()
    r['container_2'] = jar_after.relative_to(msr_dir).__str__()
    
    r['class_1'] = '/' + it['packageIdBefore'].replace('.', '/') + '/' + it['changedClass'] + '.class'
    r['class_2'] = '/' + it['packageIdAfter'].replace('.', '/') + '/' + it['changedClass'] + '.class'
    
    r['cve_id'] = it['cveId']
    r['ghsa_id'] = it['ghsaId']
    r['group_artifact'] = it['packageName']
    r['latest_vulnerable_version'] = it['latestVulnVersion']
    r['fixed_version'] = it['fixedVersion']
    r['repo'] = it['fixCommitRepo']
    r['fix_commit'] = it['fixCommitHash']
    r['modified_methods'] = it['changedMethods']
    
    df_msr.append(r)


df_msr: pd.DataFrame = pd.DataFrame(df_msr)

df_msr.to_csv(msr_dir / 'NEQ3.tsv', sep='\t', index=False)
